Emacs: Test for turning off indentation of messages in threads
In Emacs, notmuch indents messages according to their thread
depth. This leads to unused space and sometimes strange
formatting of paragraphs with long lines. Therefore I would like
to turn it off and have the message thread look more like a
conversation, as on GMail for example.

This commit contains a test for such functionality. It tests
notmuch-show for a message thread without the indentation to
indicate thread depth of messages.